:boom: Refine Major Page
