--- portal-jtk/src/lib/data/examples/Major/brief.ts
@@ -1,28 +1,0 @@
-import type { Major } from "../../../types/Major/brief.type";
-
-const majors: Array<Major> = [
-  {
-    name: "D-III Program Studi Teknik Informatika",
-    description: "D-3 bidang teknik komputer dan informatika telah berdiri sejak 1977 dan merupakan pendidikan diploma bidang teknologi informasi pertama di Indonesia.",
-    vision: "Pada tahun 2020, menjadi program studi unggulan pada jenjang pendidikan diploma IV bidang Teknik Informatika ditingkat nasional bahkan internasional. Menjadi program studi yang menghasilkan tenaga profesional pengembang perangkat lunak yang kompeten, memiliki semangat terus berkembang, bermoral, berjiwa wirausaha, dan berwawasan lingkungan.",
-    mission: "<ol type=\"1\"><li>Menyelenggarakan program pendidikan yang  menghasilkan sarjana terapan di bidang Rekayasa Perangkat Lunak (RPL) yang kompeten, bermoral, berjiwa wirausaha, berwawasan lingkungan serta mampu bersaing di tingkat nasional maupun internasional</li><li>Melaksanakan penelitian dibidang Teknologi Informasi yang dapat diterapkan pada Industri atau masyarakat ditingkat nasional maupun internasional</li><li>Melakukan pengabdian kepada masyarakat melalui diseminasi ilmu dan penerapannya di lingkup nasional.</li></ol>",
-    objective: "Tujuan D4 Teknik Informatika adalah menjalankan tridharma secara konsisten dan utuh sehingga menghasilkan lulusan yang mampu: <ul><li>Menunjukkan keunggulan keahlian dan pengetahuan, serta memiliki sikap profesionalisme yang dibutuhkan untuk menjadi seorang software designer dan analyst.</li><li>Bekerja secara individu dan menjadi bagian dari suatu team untuk membangun, menyajikan dan memelihara perangkat lunak yang berkualitas.</li><li>Mengelola proyek pembangunan perangkat lunak</li></ul>",
-    qualification: "Dosen program studi ini berkualifikasi S2 dan S3  dari perguruan tinggi ternama dari dalam negeri (ITB, UI, Tel-U, dll.), maupun dari luar negeri (Inggris, Jepang, Amerika, Australia, dan Taiwan). Mayoritas dosen telah tersertifikasi (sertifikasi dosen).",
-    accreditation: "A",
-    slug: "d3-teknik-informatika",
-    updatedAt: "2022-05-20T00:46:37.653Z"
-  },
-  {
-    name: "D-IV Program Studi Teknik Informatika",
-    description: "Program pendidikan D IV (sarjana terapan) bidang informatika di Jurusan Teknik Komputer dan Informatika didirikan pada bulan juli 2009 untuk menjawab kebutuhan terhadap penyediaan SDM yang berkaitan dengan produksi dan pemanfaatan perangkat lunak. Program studi ini dikukuhkan melalui SK Penyelenggaraan Program Studi : dari Dikti dengan nomor 1265/D/T/2009.",
-    vision: "Pada tahun 2020, menjadi program studi unggulan pada jenjang pendidikan diploma IV bidang Teknik Informatika ditingkat nasional bahkan internasional. Menjadi program studi yang menghasilkan tenaga profesional pengembang perangkat lunak yang kompeten, memiliki semangat terus berkembang, bermoral, berjiwa wirausaha, dan berwawasan lingkungan.",
-    mission: "<ol type=\"1\"><li>Menyelenggarakan program pendidikan yang  menghasilkan sarjana terapan di bidang Rekayasa Perangkat Lunak (RPL) yang kompeten, bermoral, berjiwa wirausaha, berwawasan lingkungan serta mampu bersaing di tingkat nasional maupun internasional</li><li>Melaksanakan penelitian dibidang Teknologi Informasi yang dapat diterapkan pada Industri atau masyarakat ditingkat nasional maupun internasional</li><li>Melakukan pengabdian kepada masyarakat melalui diseminasi ilmu dan penerapannya di lingkup nasional.</li></ol>",
-    objective: "Tujuan D4 Teknik Informatika adalah menjalankan tridharma secara konsisten dan utuh sehingga menghasilkan lulusan yang mampu: <ul><li>Menunjukkan keunggulan keahlian dan pengetahuan, serta memiliki sikap profesionalisme yang dibutuhkan untuk menjadi seorang software designer dan analyst.</li><li>Bekerja secara individu dan menjadi bagian dari suatu team untuk membangun, menyajikan dan memelihara perangkat lunak yang berkualitas.</li><li>Mengelola proyek pembangunan perangkat lunak</li></ul>",
-    qualification: "Dosen program studi ini berkualifikasi S2 dan S3  dari perguruan tinggi ternama dari dalam negeri (ITB, UI, Tel-U, dll.), maupun dari luar negeri (Inggris, Jepang, Amerika, Australia, dan Taiwan). Mayoritas dosen telah tersertifikasi (sertifikasi dosen).",
-    accreditation: "A",
-    slug: "d4-teknik-informatika",
-    updatedAt: "2022-05-20T00:46:37.653Z"
-  },
-];
-
-export default majors;
