+++ portal-jtk/src/lib/data/examples/Major/major.ts
@@ -1,0 +1,52 @@
+import type { Major } from "../../../types/Major/major.type";
+
+const majors: Array<Major> = [
+  {
+    name: "D-III Program Studi Teknik Informatika",
+    description:
+      "D-3 bidang teknik komputer dan informatika telah berdiri sejak 1977 dan merupakan pendidikan diploma bidang teknologi informasi pertama di Indonesia.",
+    sections: [
+      {
+        title: "Visi dan Misi",
+        description:
+          "<b>Menjadi Program Studi unggulan dan terdepan dalam program pendidikan Diploma III Teknik Informatika yang diakui baik di tingkat nasional maupun internasional pada tahun 2025.</b><br /><br />Visi ini dijabarkan ke dalam misi sebagai berikut:<ol type='i'><li>Menyelenggarakan program pendidikan Diploma III Teknik Informatika yang diakui baik di tingkat nasional maupun internasional.</li><li>Melakukan penelitian terapan yang dapat digunakan oleh Industri, Institusi atau Masyarakat baik ditingkat nasional maupun internasional</li><li>Melaksanakan pengabdian kepada masyarakat pada lingkup lokal, regional dan nasional.</li></ol>",
+      },
+      {
+        title: "Tujuan Program Studi",
+        description: "Tujuan D3 Teknik Informatika adalah menjalankan tridharma secara konsisten dan utuh sehingga menghasilkan lulusan yang mampu:<ol type='1'><li>Menghasilkan tenaga di bidang perancangan dan implementasi perangkat lunak bisnis serta perancangan solusi bisnis berbasis teknologi informasi untuk menunjang kebutuhan masyarakat dan industri di lingkup nasional dan internasional, yang memiliki sikap dan kemampuan sebagai berikut:</li><ul><li>Beradaptasi terhadap perkembangan teknologi informasi</li><li>Belajar sepanjang hayat dan ulet</li><li>Berpikir kreatif, analitis dan sistematis</li><li>Berwirausaha</li><li>Bermoral</li><li>Berkomunikasi dalam bahasa Internasional.</li></ul><li>Menghasilkan lulusan dengan kompetensi yang diakui pada tingkat nasional maupun internasional.</li><li>Mendorong mahasiswa untuk menghasilkan produk terapan di bidang perangkat lunak bisnis yang bermanfaat bagi masyarakat dan industri baik nasional maupun internasional.</li><li>Menghasilkan produk pelayanan dan produk penelitian terapan di bidang teknologi informasi yang bermanfaat bagi masyarakat dan industri baik di tingkat nasional maupun internasional.</li></ol>",
+      },
+      {
+        title: "Kualifikasi Dosen",
+        description: "Dosen program studi ini berkualifikasi S2 dan S3  dari perguruan tinggi ternama dari dalam negeri (ITB, UI, Tel-U, dll.), maupun dari luar negeri (Inggris, Jepang, Amerika, Australia, dan Taiwan). Mayoritas dosen telah tersertifikasi (sertifikasi dosen).",
+      },
+    ],
+    accreditation: "A",
+    slug: "d3-teknik-informatika",
+    updatedAt: "2022-05-20T00:46:37.653Z",
+  },
+  {
+    name: "D-IV Program Studi Teknik Informatika",
+    description:
+      "Program pendidikan D IV (sarjana terapan) bidang informatika  di Jurusan Teknik Komputer dan Informatika  didirikan pada bulan juli 2009 untuk menjawab kebutuhan terhadap penyediaan SDM yang berkaitan dengan produksi dan pemanfaatan perangkat lunak. Program studi ini dikukuhkan melalui SK Penyelenggaraan  Program Studi : dari Dikti dengan nomor 1265/D/T/2009.",
+    sections: [
+      {
+        title: "Visi dan Misi",
+        description:
+          "<b>Menjadi program studi yang menghasilkan tenaga profesional pengembang perangkat lunak yang kompeten, memiliki semangat terus berkembang, bermoral, berjiwa wirausaha, dan berwawasan lingkungan.</b><br /><br />Visi ini dijabarkan ke dalam misi sebagai berikut:<ol type='i'><li>Menyelenggarakan program pendidikan yang  menghasilkan sarjana terapan di bidang Rekayasa Perangkat Lunak (RPL) yang kompeten, bermoral, berjiwa wirausaha, berwawasan lingkungan serta mampu bersaing di tingkat nasional maupun internasional</li><li>Melaksanakan penelitian dibidang Teknologi Informasi yang dapat diterapkan pada Industri atau masyarakat ditingkat nasional maupun internasional</li><li>Melakukan pengabdian kepada masyarakat melalui diseminasi ilmu dan penerapannya di lingkup nasional</li></ol>",
+      },
+      {
+        title: "Tujuan Program Studi",
+        description: "Tujuan D4 Teknik Informatika adalah menjalankan tridharma secara konsisten dan utuh sehingga menghasilkan lulusan yang mampu:<ul><li>Menunjukkan keunggulan keahlian dan pengetahuan, serta memiliki sikap profesionalisme yang dibutuhkan untuk menjadi seorang software designer dan analyst.</li><li>Bekerja secara individu dan menjadi bagian dari suatu team untuk membangun, menyajikan dan memelihara perangkat lunak yang berkualitas.</li><li>Mengelola proyek pembangunan perangkat lunak</li></ul>",
+      },
+      {
+        title: "Kualifikasi Dosen",
+        description: "Dosen program studi ini berkualifikasi S2 dan S3  dari perguruan tinggi ternama dari dalam negeri (ITB, UI, Tel-U, dll.), maupun dari luar negeri (Inggris, Jepang, Amerika, Australia, dan Taiwan). Mayoritas dosen telah tersertifikasi (sertifikasi dosen).",
+      },
+    ],
+    accreditation: "A",
+    slug: "d4-teknik-informatika",
+    updatedAt: "2022-05-20T00:46:37.653Z",
+  },
+];
+
+export default majors;
