--- portal-jtk/src/lib/types/About/about.type.ts
@@ -1,6 +1,0 @@
-export type AboutSection = {
-  title: string;
-  description: string;
-  detail?: string;
-  slug?: string;
-};
