+++ portal-jtk/src/lib/types/Section/section.type.ts
@@ -1,0 +1,4 @@
+export type Section = {
+  title: string;
+  description: string;
+};
